Fix bugs during re-install

+++ 1_install_helpers.sh
@@ -4,9 +4,11 @@ if [ ! -d ${HOME}/bin ] ; then
     mkdir ${HOME}/bin
 fi
 
-pushd ${HOME}
-    curl -L https://github.com/kubernetes/kompose/releases/download/v1.16.0/kompose-linux-amd64 -o kompose
-    chmod 700 ./kompose
+pushd ${HOME}/bin
+    if [ ! -f ./kompose ] ; then
+        curl -L https://github.com/kubernetes/kompose/releases/download/v1.16.0/kompose-linux-amd64 -o kompose
+        chmod 700 ./kompose
+    fi
 popd
 
 cat ${HOME}/.bashrc | grep '# Add kubernetes autocompletion' &> /dev/null
